requirements.txt
streamlit>=1.37
pandas>=2.2
gspread>=6.0
google-auth>=2.35
